4.21

+++ controllers/blogs.js
@@ -52,11 +52,24 @@ blogsRouter.post('/', async (request, response) => {
     
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-      response.status(204).end()
-})
+blogsRouter.delete('/:id', async (request, response) => { 
+    const decodedToken = await jwt.verify(request.token, process.env.SECRET)
+    if(!request.token || !decodedToken.id) {
+      response.status(400).send({error: 'bad username or password'})
+      return
+    } 
 
+    const findBlog = await Blog.findById(request.params.id)
+
+    await Blog.findByIdAndRemove(request.params.id)
+      
+      if(findBlog.user.toString() === decodedToken.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(400).end()
+    }
+})
 
 blogsRouter.put('/:id', (request, response, next) => {
   const body = request.body
